requirements.txt
conan
pandas
numpy
matplotlib
jupyterlab
cython
pybind11
clang
jinja2
